requirements.txt
python-dotenv
jira
sentence-transformers

openai>=1.37.0
httpx==0.27.2

fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
boto3==1.34.144
qdrant-client==1.9.1
PyMuPDF==1.24.7
python-docx==1.1.2
requests==2.32.3
alembic